requirements.txt
requests
pyyaml
psycopg==3.2.4
psycopg_pool==3.2.4
python-dotenv==1.0.1
more_itertools 
langchain_community
torch
nltk
langchain_anthropic
langchain_openai
ragas
rich
matplotlib